+++ ui/packages/app/web/tailwind.config.js
@@ -1,4 +1,5 @@
 const defaultTheme = require('tailwindcss/defaultTheme');
+const typography = require('@tailwindcss/typography');
 
 module.exports = {
   content: [
@@ -25,5 +26,5 @@ module.exports = {
     },
   },
   variants: {},
-  plugins: [],
+  plugins: [typography],
 };